% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_link_icon}
\alias{add_link_icon}
\title{Add a link and icon}
\usage{
add_link_icon(url_path, url_text, icon_class)
}
\arguments{
\item{url_path}{string, URL (web address)}

\item{url_text}{string, URL link text to display}

\item{icon_class}{string, font awesome icon class}
}
\value{
html URL link with icon and text
}
\description{
Writes html for a URL link to R Markdown with FontAwesome icon
}
\examples{
add_link_icon("https://www.cvmedlab.org", "cvmedlab", "fas fa-globe")

}
