% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pub_list}
\alias{pub_list}
\title{Generate a publication list from a bib file}
\usage{
pub_list(bib, pdf_dir, base_url_to_pdfs)
}
\arguments{
\item{bib}{a .bib file}

\item{pdf_dir}{relative path to directory containing pdfs}

\item{base_url_to_pdfs}{url to path containing pdfs}

\item{yml}{a formatted .yml file see example}
}
\value{
publication list with links
}
\description{
This function accepts a .bib file as input, along with a specially formatted .yml file and embeds a publication list into an R Markdown document.
}
\examples{
\dontrun{
pub_list("lab_pubs.bib",
"lab_pubs.yml",
"lab_pubs/files",
"https://www.cvmedlab.org/lab_pubs/files")
}
}
